worker_node/pyproject.toml
[tool.poetry]
name = "worker-node"
version = "0.1.0"
description = "worker node"
authors = ["Bishwo <bishwo-bijaya.shah@stud.uni-bamberg.de>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
tl-dockerwrapper = "^1.3.5"
tl-crmdtos = "^1.0.0"
tl-utilities = ">=3.0.4"
psutil = "^6.0.0"
flask = "^3.0.3"
jsonpickle = "^3.3.0"
pyyaml = "^6.0.2"
python-on-whales = "^0.73.0"
paho-mqtt = "==1.6.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
colorlog = "^6.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
